+++ Miners/XmrStakMiner.cs
@@ -33,7 +33,7 @@ namespace CryptonightProfitSwitcher.Miners
                 {
                     port = settings.XmrStakApiPort;
                 }
-                var xmrJson = Helpers.GetJsonFromUrl($"http://127.0.0.1:{port}/api.json", settings, appRootFolder);
+                var xmrJson = Helpers.GetJsonFromUrl($"http://127.0.0.1:{port}/api.json", settings, appRootFolder, CancellationToken.None);
                 dynamic xmr = JObject.Parse(xmrJson);
                 JArray totalHashRates = xmr.hashrate.total;
                 return totalHashRates[0].ToObject<double>();
@@ -53,6 +53,10 @@ namespace CryptonightProfitSwitcher.Miners
             string xmrFolderPath = Path.GetDirectoryName(xmrPath);
             var xmrDirectory = new DirectoryInfo(xmrFolderPath);
             _process.StartInfo.FileName = xmrPath;
+            if (Helpers.IsLinux())
+            {
+                _process.StartInfo.FileName = "x-terminal-emulator";
+            }
             string configPath = Helpers.ResolveToFullPath(mineable.ConfigPath, appRoot);
             File.Copy(configPath, Path.Combine(xmrFolderPath, "current_config.txt"), true);
             string args = "-c current_config.txt";
@@ -105,11 +109,16 @@ namespace CryptonightProfitSwitcher.Miners
             }
 
             _process.StartInfo.Arguments = args;
+            if (Helpers.IsLinux())
+            {
+                _process.StartInfo.Arguments = $"-e \"'{xmrPath}' {args}\"";
+            }
             _process.StartInfo.UseShellExecute = true;
             _process.StartInfo.CreateNoWindow = false;
             _process.StartInfo.RedirectStandardOutput = false;
             _process.StartInfo.WorkingDirectory = xmrFolderPath;
             _process.StartInfo.WindowStyle = settings.StartMinerMinimized ? ProcessWindowStyle.Minimized : ProcessWindowStyle.Normal;
+            _process.EnableRaisingEvents = true;
             Thread.Sleep(TimeSpan.FromSeconds(settings.MinerStartDelay));
             _process.Start();
         }
@@ -120,14 +129,49 @@ namespace CryptonightProfitSwitcher.Miners
             {
                 try
                 {
-                    _process.Kill();
+                    _process.CloseMainWindow();
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine("Couldn't close miner process: " + ex.Message);
+                }
+                try
+                {
+                    if (!_process.HasExited)
+                    {
+                        _process.Kill();
+                    }
                 }
                 catch (Exception ex)
                 {
                     Console.WriteLine("Couldn't kill miner process: " + ex.Message);
                 }
+
                 _process.Dispose();
                 _process = null;
+
+                if (Helpers.IsLinux())
+                {
+                    try
+                    {
+                        var killProcess = new Process();
+                        killProcess.StartInfo.FileName = "x-terminal-emulator";
+                        string xmrName = Path.GetFileName(_mineable.XmrStakPath);
+                        Console.WriteLine("Kill: " + xmrName);
+                        killProcess.StartInfo.Arguments = $"-e \"killall -9 {xmrName}\"";
+                        killProcess.StartInfo.UseShellExecute = true;
+                        killProcess.StartInfo.CreateNoWindow = false;
+                        killProcess.StartInfo.RedirectStandardOutput = false;
+                        killProcess.Start();
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine("Couldn't kill process: " + ex.Message);
+
+                    }
+                }
+
+
                 _mineable = null;
             }
         }
